+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [3.4.0] - 09.09.2026
 
 ### Added